requirements.txt.txt
mysql-connector-python
python-dotenv
python-telegram-bot
